+++ src/main/java/com/example/toanyone/domain/cart/controller/CartController.java
@@ -4,8 +4,8 @@ package com.example.toanyone.domain.cart.controller;
 import com.example.toanyone.domain.cart.dto.CartRequestDto;
 import com.example.toanyone.domain.cart.dto.CartResponseDto;
 import com.example.toanyone.domain.cart.service.CartService;
-import com.example.toanyone.global.common.annotation.Auth;
-import com.example.toanyone.global.common.dto.AuthUser;
+import com.example.toanyone.global.auth.annotation.Auth;
+import com.example.toanyone.global.auth.dto.AuthUser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
